telegram_bot/pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = "Connect TelegramBot"
authors = ["kiril <kirill.301522@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.4.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
